pyproject.toml
[project]
name = "pothole-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.1",
    "python-dotenv>=1.2.2",
    "roboflow>=1.3.7",
    "ultralytics>=8.4.45",
    "uvicorn>=0.46.0",
]
